changed Deal.GetValue to return decimal as it is a financial value (https://stackoverflow.com/a/5650441/2999220)

+++ ZendeskSell/Deals/DealRequest.cs
@@ -13,7 +13,7 @@ namespace ZendeskSell.Deals {
 
         [JsonProperty("value")]
         public string Value { get; set; }
-        public float GetValue() => float.Parse(Value, System.Globalization.NumberFormatInfo.InvariantInfo);
+        public decimal GetValue() => decimal.Parse(Value, System.Globalization.NumberFormatInfo.InvariantInfo);
 
         [JsonProperty("name")]
         public string Name { get; set; }
